src/requirements.txt
injector==0.20.1
click==8.1.3
environs==9.5.0
jsonpickle==2.2.0
pytest
pytest-cov
asabix-python-sqlalchemy-connection==1.4.2
asabix_python_utility_commons_lib==1.6.4
